[BPK-1396] Remove central storybook dectorator
