+++ native/packages/react-native-bpk-component-horizontal-nav/stories.js
@@ -27,6 +27,7 @@ import { spacingBase } from 'bpk-tokens/tokens/base.react.native';
 import BpkHorizontalNav, { BpkHorizontalNavItem } from './index';
 import { StorySubheading } from '../../storybook/TextStyles';
 import themeAttributes from '../../storybook/themeAttributes';
+import CenterDecorator from '../../storybook/CenterDecorator';
 
 const styles = StyleSheet.create({
   bottomMargin: {
@@ -97,6 +98,7 @@ StoryNav.defaultProps = {
 };
 
 storiesOf('BpkHorizontalNav', module)
+  .addDecorator(CenterDecorator)
   .add('docs:default', () => (
     <View style={styles.bottomMargin}>
       <BpkHorizontalNav selectedId="1">
