+++ native/packages/react-native-bpk-component-star-rating/stories.js
@@ -21,6 +21,7 @@ import { View, StyleSheet, Text } from 'react-native';
 import { storiesOf } from '@storybook/react-native';
 import { spacingBase } from 'bpk-tokens/tokens/base.react.native';
 import BpkStarRating, { BpkStar, STAR_TYPES } from './index';
+import CenterDecorator from '../../storybook/CenterDecorator';
 
 const StarRating = props => (
   <BpkStarRating ratingLabel={(r, m) => `${r} out of ${m} stars`} {...props} />
@@ -31,6 +32,7 @@ const styles = StyleSheet.create({
   },
 });
 storiesOf('BpkStarRating', module)
+  .addDecorator(CenterDecorator)
   .add('docs:default', () => (
     <View>
       <View style={styles.container}>
@@ -85,6 +87,7 @@ storiesOf('BpkStarRating', module)
   ));
 
 storiesOf('BpkStar', module)
+  .addDecorator(CenterDecorator)
   .add('Full', () => (
     <View style={styles.container}>
       <Text>Full Star</Text>
