+++ native/packages/react-native-bpk-component-touchable-native-feedback/stories.js
@@ -20,6 +20,7 @@ import React from 'react';
 import { View, StyleSheet } from 'react-native';
 import BpkText from 'react-native-bpk-component-text';
 import { storiesOf, action } from '@storybook/react-native';
+import CenterDecorator from '../../storybook/CenterDecorator';
 
 import BpkTouchableNativeFeedback from './index';
 
@@ -44,6 +45,7 @@ const viewContent = <View style={styles.view}>{textContent}</View>;
 const onPress = action('Touchable native feedback press');
 
 storiesOf('BpkTouchableNativeFeedback', module)
+  .addDecorator(CenterDecorator)
   .add('docs:text-content', () => (
     <BpkTouchableNativeFeedback
       onPress={onPress}
